webapp/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
click==8.1.3
Django==4.0.6
django-rest-framework==0.1.0
djangorestframework==3.13.1
htcondor==9.4.0
joblib==1.1.0
nltk==3.7
numpy==1.23.1
pandas==1.4.3
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
regex==2022.7.9
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
